Update ID types to long and change date handling
- Changed ID parameters in controllers from int to long.
- Updated Categoria, TipoTransacao, and Transacao models to use long for IDs.
- Modified database migrations to reflect bigint types for ID columns.
- Updated AppDbContextModelSnapshot for consistency with new ID types.
- Created new migrations for initial database setup and data population.
- Changed Transacao model's Data property from DateTime to DateOnly.
- Added migration to adjust Data column in TRA_transacao to DateOnly.

+++ FintrackAPI/Models/Categoria.cs
@@ -15,7 +15,7 @@ public class Categoria
 
     [Key]
     [Column("cat_id_categoria")]
-    public int CategoriaId { get; set; }
+    public long CategoriaId { get; set; }
 
     [Required]
     [StringLength(50)]
